// admin/src/Services/Banner/interfaces.ts
import {
	VinistoHelperDllEnumsSliderCarouselButtonStyle,
	VinistoHelperDllEnumsSliderCarouselCarouselType,
} from 'vinisto_api_client/src/api-types/cms-api/';

import { POSITION } from './constants';

export interface BannerApi {
	id?: string;
	type: VinistoHelperDllEnumsSliderCarouselCarouselType;
	title?: string;
	subtitle?: string;
	textLink?: string;
	availableFrom: number;
	availableTo?: number;
	url: string;
	position: number;
}

export interface BannerImage {
	id: string;
	url: string;
}

// TODO Get rid of this, use autogenerated type
export interface Banner {
	id: string;
	position: POSITION;
	title: string;
	titleColor?: string;
	subtitle: string;
	subtitleColor?: string;
	image: BannerImage;
	ctaLabel: string;
	buttonStyle: VinistoHelperDllEnumsSliderCarouselButtonStyle;
	validFrom: Date;
	validTo: Date;
	order: number | null;
	url: string;
	availableOnPlatforms: number[];
}